+++ inst/tinytest/test-tt.R
@@ -9,3 +9,12 @@ expect_equal(length(colnames(tt(df1))), 0)
 tab1 = save_tt(tt(df1), "markdown")
 tab2 = save_tt(tt(df2, colnames = FALSE), "markdown")
 expect_equal(tab1, tab2)
+
+
+# colnames="label"
+dat <- mtcars[1:5, c("cyl", "mpg", "hp")]
+attr(dat$cyl, "label") <- "Cylinders"
+attr(dat$mpg, "label") <- "Miles per Gallon"
+attr(dat$hp, "label") <- "Horse Power"
+tab <- tt(dat, colnames = "label") |> save_tt("dataframe")
+expect_equal(unlist(tab[1, ]), c("Cylinders", "Miles per Gallon", "Horse Power"))